feat: Cli version to fix the template
