+++ run.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python
 
-from handler import Handler
-from http.server import HTTPServer
+# from server import run
+from cli import run
 
-with HTTPServer(('', 8000), Handler) as server:
-    server.serve_forever()
+run()
